reverting to Nick's r1008 btree deletes and inserts, with an added bug fix
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@1063 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ lib/btree/btree_utils.h
@@ -36,13 +36,13 @@ void btree_unlock(struct btree *tree) ;
 #define BTREE_OP_ATTR	static __inline __attribute__((always_inline))
 
 BTREE_OP_ATTR const offset_t *
-btree_key(const offset_t *keys, const int i, uint8_t s_key)
+btree_key(const offset_t *keys, const int i, int s_key)
 {
     return &keys[i * s_key] ;
 }
 
 BTREE_OP_ATTR int64_t
-btree_keycmp(const offset_t *key1, const offset_t *key2, uint8_t s_key)
+btree_keycmp(const offset_t *key1, const offset_t *key2, int s_key)
 {
     int i = 0 ; 
     int64_t r = 0 ;
@@ -52,19 +52,19 @@ btree_keycmp(const offset_t *key1, const offset_t *key2, uint8_t s_key)
 }
 
 BTREE_OP_ATTR void
-btree_keycpy(const offset_t *dst, const offset_t *src, uint8_t s_key)
+btree_keycpy(const offset_t *dst, const offset_t *src, int s_key)
 {
     memcpy((offset_t *)dst, src, s_key * sizeof(offset_t)) ;
 }
 
 BTREE_OP_ATTR void
-btree_keymove(const offset_t *dst, const offset_t *src, uint8_t s_key)
+btree_keymove(const offset_t *dst, const offset_t *src, int s_key)
 {
     memmove((offset_t *)dst, src, s_key * sizeof(offset_t)) ;
 }
 
 BTREE_OP_ATTR void
-btree_keyset(const offset_t *dst, offset_t val, uint8_t s_key)
+btree_keyset(const offset_t *dst, offset_t val, int s_key)
 {
     offset_t *d = (offset_t *) dst;
     
@@ -76,25 +76,25 @@ btree_keyset(const offset_t *dst, offset_t val, uint8_t s_key)
 }
 
 BTREE_OP_ATTR const offset_t *
-btree_value(const offset_t *vals, const int i, uint8_t s_val)
+btree_value(const offset_t *vals, const int i, int s_val)
 {
     return &vals[i * s_val] ;
 }
 
 BTREE_OP_ATTR void
-btree_valcpy(const offset_t *dst, const offset_t *src, uint8_t s_val)
+btree_valcpy(const offset_t *dst, const offset_t *src, int s_val)
 {
     memcpy((offset_t *)dst, src, s_val * sizeof(offset_t)) ;
 }
 
 BTREE_OP_ATTR void
-btree_valmove(const offset_t *dst, const offset_t *src, uint8_t s_val)
+btree_valmove(const offset_t *dst, const offset_t *src, int s_val)
 {
     memmove((offset_t *)dst, src, s_val * sizeof(offset_t)) ;
 }
 
 BTREE_OP_ATTR void
-btree_valset(const offset_t *dst, offset_t val, uint8_t s_val)
+btree_valset(const offset_t *dst, offset_t val, int s_val)
 {
     offset_t *d = (offset_t *) dst;
